changed type Double for price values to BigDecimal; added changelog related to changes with price type;

+++ src/main/java/com/gp/beershop/mock/BeerMock.java
@@ -2,6 +2,7 @@ package com.gp.beershop.mock;
 
 import com.gp.beershop.dto.Beer;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -19,7 +20,7 @@ public final class BeerMock {
             .alcohol(5.0)
             .density(11.5)
             .country("Республика Беларусь")
-            .price(5D).build());
+            .price(BigDecimal.valueOf(5D)).build());
         put(2, Beer.builder()
             .id(2)
             .type("темное")
@@ -29,7 +30,7 @@ public final class BeerMock {
             .alcohol(4.6)
             .density(10.2)
             .country("Республика Беларусь")
-            .price(3D).build());
+            .price(BigDecimal.valueOf(3D)).build());
         put(3, Beer.builder()
             .id(3)
             .type("светлое осветлённое")
@@ -39,7 +40,7 @@ public final class BeerMock {
             .alcohol(4.2)
             .density(12.0)
             .country("Чехия")
-            .price(8D).build());
+            .price(BigDecimal.valueOf(8D)).build());
         put(4, Beer.builder()
             .id(3)
             .type("светлое осветлённое")
@@ -49,7 +50,7 @@ public final class BeerMock {
             .alcohol(4.5)
             .density(12.2)
             .country("Чехия")
-            .price(8.3).build());
+            .price(BigDecimal.valueOf(8.3)).build());
     }};
 
     public static Beer getById(final Integer id) {
